refactored snake game table into its own component
